move additional_input_name_for_pipeline_channel
