+++ sdk/python/kfp/compiler/pipeline_spec_builder_test.py
@@ -19,7 +19,6 @@ from absl.testing import parameterized
 from google.protobuf import json_format
 from google.protobuf import struct_pb2
 from kfp.compiler import pipeline_spec_builder
-from kfp.components import pipeline_channel
 from kfp.pipeline_spec import pipeline_spec_pb2
 import yaml
 
@@ -28,39 +27,6 @@ class PipelineSpecBuilderTest(parameterized.TestCase):
 
     def setUp(self):
         self.maxDiff = None
-
-    @parameterized.parameters(
-        {
-            'channel':
-                pipeline_channel.PipelineParameterChannel(
-                    name='output1', task_name='task1', channel_type='String'),
-            'expected':
-                'pipelinechannel--task1-output1',
-        },
-        {
-            'channel':
-                pipeline_channel.PipelineArtifactChannel(
-                    name='output1',
-                    task_name='task1',
-                    channel_type='system.Artifact@0.0.1',
-                ),
-            'expected':
-                'pipelinechannel--task1-output1',
-        },
-        {
-            'channel':
-                pipeline_channel.PipelineParameterChannel(
-                    name='param1', channel_type='String'),
-            'expected':
-                'pipelinechannel--param1',
-        },
-    )
-    def test_additional_input_name_for_pipeline_channel(self, channel,
-                                                        expected):
-        self.assertEqual(
-            expected,
-            pipeline_spec_builder.additional_input_name_for_pipeline_channel(
-                channel))
 
     @parameterized.parameters(
         {
